+++ client/src/Components/community/SearchPlace.tsx
