+++ client/src/Components/mainpage/Banner.tsx
