+++ client/src/Components/mainpage/CarouselNearbyPlace.tsx
