+++ client/src/Components/member/IntroBox.tsx
