+++ client/src/pages/Community/PostUpdate.tsx
@@ -13,8 +13,8 @@ import colorSyntax from '@toast-ui/editor-plugin-color-syntax';
 /* eslint-disable jsx-a11y/label-has-associated-control */
 import { Editor } from '@toast-ui/react-editor';
 
-import SearchPlace from '../../ccc/community/SearchPlace';
-import SubjectDropdown from '../../ccc/community/SubjectDropdown';
+import SearchPlace from '../../Components/community/SearchPlace';
+import SubjectDropdown from '../../Components/community/SubjectDropdown';
 import { Api } from '../../apis/customAPI';
 import useAxios from '../../hooks/useAxios';
